Moving helper functions to respective page test files.
