+++ frontend/e2e/protected/apply/adult-category.spec.ts
@@ -2,7 +2,7 @@ import { test } from '@playwright/test';
 
 import { AdultPage } from '../../pages/protected/apply/adult-page';
 import { InitialPage } from '../../pages/protected/apply/initial-page';
-import { acceptLegalCheckboxes, calculateDOB, clickContinue, fillApplicantInformationForm, fillOutAddress } from '../../utils/helpers';
+import { calculateDOB, clickContinue } from '../../utils/helpers';
 
 test.describe('Adult category', () => {
   test.beforeEach('Navigate to adult application', async ({ page }) => {
@@ -13,7 +13,7 @@ test.describe('Adult category', () => {
 
     // Accept Terms
     await applyPage.isLoaded('terms-and-conditions');
-    await acceptLegalCheckboxes(page);
+    await applyPage.acceptLegalCheckboxes(page);
     await clickContinue(page);
 
     // Tax Filing Section
@@ -34,7 +34,7 @@ test.describe('Adult category', () => {
     await test.step('Should navigate to applicant information page', async () => {
       await applyAdultPage.isLoaded('applicant-information');
       const { year, month, day } = calculateDOB(35);
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
+      await applyAdultPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
 
       // DTC question
       await page.getByRole('radio', { name: 'Yes', exact: true }).check();
@@ -55,14 +55,14 @@ test.describe('Adult category', () => {
 
     await test.step('Should navigate to mailing address page', async () => {
       await applyAdultPage.isLoaded('mailing-address');
-      await fillOutAddress({ address: '123 Fake Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
+      await applyAdultPage.fillOutAddress({ address: '123 Fake Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
       await clickContinue(page);
       await page.getByRole('button', { name: 'Use selected address' }).click();
     });
 
     await test.step('Should navigate to home address page', async () => {
       await applyAdultPage.isLoaded('home-address');
-      await fillOutAddress({ address: '123 Maple Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
+      await applyAdultPage.fillOutAddress({ address: '123 Maple Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
       await clickContinue(page);
       await page.getByRole('button', { name: 'Use selected address' }).click();
     });
@@ -139,7 +139,7 @@ test.describe('Adult category', () => {
     await test.step('Should navigate to applicant information page', async () => {
       await applyAdultPage.isLoaded('applicant-information');
       const { year, month, day } = calculateDOB(15);
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
+      await applyAdultPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
 
       await clickContinue(page);
     });
@@ -156,7 +156,7 @@ test.describe('Adult category', () => {
     await test.step('Should navigate to applicant information page', async () => {
       await applyAdultPage.isLoaded('applicant-information');
       const { year, month, day } = calculateDOB(17);
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
+      await applyAdultPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
 
       await clickContinue(page);
     });
@@ -180,7 +180,7 @@ test.describe('Adult category', () => {
     await test.step('Should navigate to applicant information page', async () => {
       await applyAdultPage.isLoaded('applicant-information');
       const { year, month, day } = calculateDOB(17);
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
+      await applyAdultPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: true, page });
 
       await clickContinue(page);
     });
@@ -211,7 +211,7 @@ test.describe('Adult category', () => {
 
     await test.step('Should navigate to applicant information page', async () => {
       await applyAdultPage.isLoaded('applicant-information');
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: '10', month: '10', year: '2006', dtcEligible: true, page });
+      await applyAdultPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: '10', month: '10', year: '2006', dtcEligible: true, page });
 
       await clickContinue(page);
     });
