+++ frontend/e2e/protected/apply/child-category.spec.ts
@@ -2,7 +2,7 @@ import { test } from '@playwright/test';
 
 import { ChildPage } from '../../pages/protected/apply/child-page';
 import { InitialPage } from '../../pages/protected/apply/initial-page';
-import { acceptLegalCheckboxes, calculateDOB, clickContinue, fillApplicantInformationForm, fillChildrenInformationForm, fillOutAddress } from '../../utils/helpers';
+import { calculateDOB, clickContinue } from '../../utils/helpers';
 
 test.describe('Child category', () => {
   test.beforeEach('Navigate to child application', async ({ page }) => {
@@ -13,7 +13,7 @@ test.describe('Child category', () => {
 
     // Accept Terms
     await applyPage.isLoaded('terms-and-conditions');
-    await acceptLegalCheckboxes(page);
+    await applyPage.acceptLegalCheckboxes(page);
     await clickContinue(page);
 
     // Tax Filing Section
@@ -40,7 +40,7 @@ test.describe('Child category', () => {
     await test.step('Should navigate to children-information page', async () => {
       await applyChildPage.isLoaded('children-information');
       const { year, month, day } = calculateDOB(10);
-      await fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', sin: '800000002', day: day, month: month, year: year, hasSin: true, isGuardian: true, page });
+      await applyChildPage.fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', sin: '800000002', day: day, month: month, year: year, hasSin: true, isGuardian: true, page });
       await clickContinue(page);
     });
 
@@ -80,7 +80,7 @@ test.describe('Child category', () => {
     await test.step('Should navigate to applicant information page', async () => {
       await applyChildPage.isLoaded('applicant-information');
       const { year, month, day } = calculateDOB(35);
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: undefined, page });
+      await applyChildPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: undefined, page });
 
       await clickContinue(page);
     });
@@ -93,14 +93,14 @@ test.describe('Child category', () => {
 
     await test.step('Should navigate to mailing address page', async () => {
       await applyChildPage.isLoaded('mailing-address');
-      await fillOutAddress({ address: '123 Fake Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
+      await applyChildPage.fillOutAddress({ address: '123 Fake Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
       await clickContinue(page);
       await page.getByRole('button', { name: 'Use selected address' }).click();
     });
 
     await test.step('Should navigate to home address page', async () => {
       await applyChildPage.isLoaded('home-address');
-      await fillOutAddress({ address: '123 Maple Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
+      await applyChildPage.fillOutAddress({ address: '123 Maple Street', city: 'Ottawa', country: 'Canada', province: 'Ontario', postalCode: 'K1A 0B1', page });
       await clickContinue(page);
       await page.getByRole('button', { name: 'Use selected address' }).click();
     });
@@ -161,7 +161,7 @@ test.describe('Child category', () => {
     await test.step('Should navigate to children-information page', async () => {
       await applyChildPage.isLoaded('children-information');
       const { year, month, day } = calculateDOB(10);
-      await fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', sin: '800000002', day: day, month: month, year: year, hasSin: true, isGuardian: false, page });
+      await applyChildPage.fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', sin: '800000002', day: day, month: month, year: year, hasSin: true, isGuardian: false, page });
       await clickContinue(page);
     });
 
@@ -183,7 +183,7 @@ test.describe('Child category', () => {
     await test.step('Should navigate to children-information page', async () => {
       await applyChildPage.isLoaded('children-information');
       const { year, month, day } = calculateDOB(18);
-      await fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', day: day, month: month, year: year, hasSin: false, isGuardian: true, page });
+      await applyChildPage.fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', day: day, month: month, year: year, hasSin: false, isGuardian: true, page });
       await clickContinue(page);
     });
 
@@ -204,7 +204,7 @@ test.describe('Child category', () => {
     await test.step('Should navigate to children-information page', async () => {
       await applyChildPage.isLoaded('children-information');
       const { year, month, day } = calculateDOB(17);
-      await fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', day: day, month: month, year: year, hasSin: false, isGuardian: true, page });
+      await applyChildPage.fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', day: day, month: month, year: year, hasSin: false, isGuardian: true, page });
       await clickContinue(page);
     });
 
@@ -225,7 +225,7 @@ test.describe('Child category - parent or legual guardian miscellaneous checks',
 
     // Accept Terms
     await applyPage.isLoaded('terms-and-conditions');
-    await acceptLegalCheckboxes(page);
+    await applyPage.acceptLegalCheckboxes(page);
     await clickContinue(page);
 
     // Tax Filing Section
@@ -244,7 +244,7 @@ test.describe('Child category - parent or legual guardian miscellaneous checks',
 
     await applyChildPage.isLoaded('children-information');
     const { year, month, day } = calculateDOB(10);
-    await fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', sin: '800000002', day: day, month: month, year: year, hasSin: true, isGuardian: true, page });
+    await applyChildPage.fillChildrenInformationForm({ firstName: 'John Jr.', lastName: 'Smith', sin: '800000002', day: day, month: month, year: year, hasSin: true, isGuardian: true, page });
     await clickContinue(page);
 
     await applyChildPage.isLoaded('children-dental-insurance');
@@ -278,7 +278,7 @@ test.describe('Child category - parent or legual guardian miscellaneous checks',
     await test.step('Should navigate to applicant information page', async () => {
       await applyChildPage.isLoaded('applicant-information');
       const { year, month, day } = calculateDOB(15);
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: undefined, page });
+      await applyChildPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: day, month: month, year: year, dtcEligible: undefined, page });
 
       await clickContinue(page);
     });
@@ -292,7 +292,7 @@ test.describe('Child category - parent or legual guardian miscellaneous checks',
     const applyChildPage = new ChildPage(page);
     await test.step('Should navigate to applicant information page', async () => {
       await applyChildPage.isLoaded('applicant-information');
-      await fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: '10', month: '10', year: '2006', dtcEligible: undefined, page });
+      await applyChildPage.fillApplicantInformationForm({ firstName: 'John', lastName: 'Smith', sin: '900000001', day: '10', month: '10', year: '2006', dtcEligible: undefined, page });
 
       await clickContinue(page);
     });
